v1.2.1 - Revision 1 - Refactored code to compile for version 1.2.1. - Removed unused import statements. - Removed internal usages of deprecated classes.
