+++ build.gradle
@@ -4,7 +4,7 @@ plugins {
 }
 
 group 'io.github.math0898'
-version 'v1.1.1-BETA'
+version 'v1.2.1'
 
 repositories {
     mavenCentral()
