+++ src/main/java/sugaEngine/sound/SoundManager.java
@@ -1,10 +1,5 @@
 package sugaEngine.sound;
 
-import java.io.IOException;
-import java.nio.ByteBuffer;
-import java.nio.IntBuffer;
-import java.util.Objects;
-
 //import org.lwjgl.openal.*;
 
 /**
